+++ kcm/autoscroll_config.cpp
@@ -12,7 +12,7 @@
 #include <QDBusMessage>
 #include <QDBusPendingCall>
 
-K_PLUGIN_CLASS(AutoScroll::AutoScrollEffectConfig)
+K_PLUGIN_CLASS_WITH_JSON(AutoScroll::AutoScrollEffectConfig, "metadata.json")
 
 namespace AutoScroll {
 
@@ -24,6 +24,11 @@ AutoScrollEffectConfig::AutoScrollEffectConfig(QObject *parent,
       KSharedConfig::openConfig(QStringLiteral("kwinrc")));
   m_config->read();
   addConfig(m_config.get(), widget());
+
+  connect(m_ui.kcfg_GlyphStyle, &GlyphStyleComboBox::glyphStyleChanged,
+          m_ui.kcfg_GlyphSize, &GlyphSizeComboBox::setGlyphStyle);
+  m_ui.kcfg_GlyphSize->setGlyphStyle(m_ui.kcfg_GlyphStyle->glyphStyle());
+  m_ui.appearanceGroupBox->setEnabled(m_ui.kcfg_VisualFeedback->isChecked());
 }
 
 AutoScrollEffectConfig::~AutoScrollEffectConfig() = default;