+++ src/autoscrollvisual.cpp
@@ -3,6 +3,8 @@
 
 #include "autoscrollvisual.h"
 
+#include "glyphstyle.h"
+
 #include <scene/imageitem.h>
 #include <scene/item.h>
 #ifdef AUTOSCROLL_USE_LEGACY_IMAGE_ITEM_FACTORY
@@ -10,15 +12,9 @@
 #include <scene/scene.h>
 #endif
 
-#include <QPainter>
-#include <QSvgRenderer>
-
-#include <cmath>
-
 namespace AutoScroll {
 namespace {
-constexpr qreal AnchorSize = 40.0;
-constexpr qreal CursorSize = 32.0;
+constexpr qreal DirectionSizeRatio = 0.8;
 
 std::unique_ptr<KWin::ImageItem> createImageItem(KWin::Item *parent) {
 #ifdef AUTOSCROLL_USE_LEGACY_IMAGE_ITEM_FACTORY
@@ -28,26 +24,6 @@ std::unique_ptr<KWin::ImageItem> createImageItem(KWin::Item *parent) {
 #endif
 }
 
-QImage renderSvg(const QString &resource, qreal logicalSize, qreal scale,
-                 qreal rotation) {
-  const int pixelSize =
-      std::max(1, static_cast<int>(std::ceil(logicalSize * scale)));
-  QImage image(QSize(pixelSize, pixelSize),
-               QImage::Format_ARGB32_Premultiplied);
-  image.fill(Qt::transparent);
-
-  QSvgRenderer renderer(resource);
-  {
-    QPainter painter(&image);
-    painter.setRenderHint(QPainter::Antialiasing);
-    painter.translate(pixelSize / 2.0, pixelSize / 2.0);
-    painter.rotate(rotation);
-    painter.translate(-pixelSize / 2.0, -pixelSize / 2.0);
-    renderer.render(&painter, QRectF(0, 0, pixelSize, pixelSize));
-  }
-  image.setDevicePixelRatio(scale);
-  return image;
-}
 } // namespace
 
 AutoScrollVisual::AutoScrollVisual(KWin::Item *overlayItem)
@@ -55,8 +31,37 @@ AutoScrollVisual::AutoScrollVisual(KWin::Item *overlayItem)
 
 AutoScrollVisual::~AutoScrollVisual() = default;
 
+void AutoScrollVisual::setAppearance(const QString &styleId, int glyphSize) {
+  const QString normalizedStyle = normalizedGlyphStyleId(styleId);
+  const qreal normalizedSize = normalizedGlyphSize(glyphSize);
+  if (m_styleId == normalizedStyle &&
+      qFuzzyCompare(m_anchorSize, normalizedSize)) {
+    return;
+  }
+
+  m_styleId = normalizedStyle;
+  m_anchorSize = normalizedSize;
+  const qreal directionSize = m_anchorSize * DirectionSizeRatio;
+
+  if (m_anchorItem) {
+    m_anchorItem->setImage(renderAnchor(m_scale));
+    m_anchorItem->setSize(QSizeF(m_anchorSize, m_anchorSize));
+    m_anchorItem->setPosition(m_anchorPosition -
+                              QPointF(m_anchorSize / 2.0, m_anchorSize / 2.0));
+  }
+  if (m_cursorItem) {
+    m_cursorItem->setSize(QSizeF(directionSize, directionSize));
+    m_cursorItem->setPosition(
+        m_cursorPosition - QPointF(directionSize / 2.0, directionSize / 2.0));
+    if (m_direction != Direction::Center) {
+      m_cursorItem->setImage(renderDirection(m_direction, m_scale));
+    }
+  }
+}
+
 void AutoScrollVisual::show(const QPointF &anchorPosition,
                             const QPointF &cursorPosition, qreal scale) {
+  hide();
   m_anchorPosition = anchorPosition;
   m_cursorPosition = cursorPosition;
   m_scale = std::max(scale, 1.0);
@@ -64,13 +69,14 @@ void AutoScrollVisual::show(const QPointF &anchorPosition,
 
   m_anchorItem = createImageItem(m_overlayItem);
   m_anchorItem->setImage(renderAnchor(m_scale));
-  m_anchorItem->setSize(QSizeF(AnchorSize, AnchorSize));
+  m_anchorItem->setSize(QSizeF(m_anchorSize, m_anchorSize));
   m_anchorItem->setPosition(m_anchorPosition -
-                            QPointF(AnchorSize / 2.0, AnchorSize / 2.0));
+                            QPointF(m_anchorSize / 2.0, m_anchorSize / 2.0));
   m_anchorItem->setZ(1000);
 
+  const qreal directionSize = m_anchorSize * DirectionSizeRatio;
   m_cursorItem = createImageItem(m_overlayItem);
-  m_cursorItem->setSize(QSizeF(CursorSize, CursorSize));
+  m_cursorItem->setSize(QSizeF(directionSize, directionSize));
   m_cursorItem->setZ(1001);
   m_cursorItem->setVisible(false);
 }
@@ -100,8 +106,9 @@ void AutoScrollVisual::update(const QPointF &cursorPosition,
   if (imageChanged) {
     m_cursorItem->setImage(renderDirection(m_direction, m_scale));
   }
+  const qreal directionSize = m_anchorSize * DirectionSizeRatio;
   m_cursorItem->setPosition(m_cursorPosition -
-                            QPointF(CursorSize / 2.0, CursorSize / 2.0));
+                            QPointF(directionSize / 2.0, directionSize / 2.0));
   m_cursorItem->setVisible(true);
 }
 
@@ -113,14 +120,14 @@ void AutoScrollVisual::hide() {
 bool AutoScrollVisual::isVisible() const { return m_anchorItem != nullptr; }
 
 QImage AutoScrollVisual::renderAnchor(qreal scale) const {
-  return renderSvg(QStringLiteral(":/autoscroll/anchor.svg"), AnchorSize, scale,
-                   0.0);
+  return renderGlyph(m_styleId, GlyphElement::Anchor, m_anchorSize, scale);
 }
 
 QImage AutoScrollVisual::renderDirection(Direction direction,
                                          qreal scale) const {
-  return renderSvg(QStringLiteral(":/autoscroll/direction.svg"), CursorSize,
-                   scale, rotationForDirection(direction));
+  return renderGlyph(m_styleId, GlyphElement::Direction,
+                     m_anchorSize * DirectionSizeRatio, scale,
+                     rotationForDirection(direction));
 }
 
 qreal AutoScrollVisual::rotationForDirection(Direction direction) {